Fix error with modifications
